Using postgres-async, but not working with date formatting - will try straight JDBC

+++ BackEnd/src/main/scala/wikimap/parser/FileEventExtractor.scala
@@ -18,7 +18,7 @@ object FileEventExtractor {
       .toList
       .flatMap(_ match {
         case SimpleEventString(date, month, year, desc) =>
-          Some(SimpleEvent(Date(date.toInt, month, year.toInt), desc))
+          Some(SimpleEvent(Date(date.toInt, month.toInt, year.toInt), desc))
         case _ =>
           None
       })
